194 Add the ability to add Content Delivery API Headers.
